# docs/openapi.yaml
openapi: 3.1.0
info:
  title: EMotorad Storefront BFF Prototype
  version: 0.1.0
  description: Contract scaffold. Prototype handlers return mock data and perform no production writes.
servers:
  - url: /api
paths:
...
  /leads:
    post:
      summary: Accept a normalized marketing lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, phone]
              properties:
                name: { type: string }
                phone: { type: string }
                source: { type: string }
                consent: { type: boolean }
      responses:
        "202": { description: Lead accepted for processing }
        "400": { $ref: "#/components/responses/BadRequest" }
...
  /test-rides/bookings:
    post:
      summary: Request a test-ride booking
      responses:
        "202": { description: Booking request accepted }
        "400": { $ref: "#/components/responses/BadRequest" }
components:
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            correlationId: { type: string }
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...